refactor ddtrace.testing.internal.logs module
